+++ common/core/src/commonMain/kotlin/database/room/dao/AuthorsDao.kt
@@ -15,8 +15,8 @@ interface AuthorsDao {
     suspend fun updateAuthor(author: AuthorEntity)
 
     @Query("SELECT * FROM AuthorEntity WHERE id = :authorId AND userId = :userId")
-    suspend fun getAuthorByAuthorId(authorId: String, userId: Long): List<AuthorEntity>
+    suspend fun getAuthorByAuthorId(authorId: String, userId: Int): List<AuthorEntity>
 
     @Query("SELECT * FROM AuthorEntity WHERE timestampOfUpdating > :timestamp AND userId = :userId")
-    suspend fun getNotSynchronizedAuthors(timestamp: Long, userId: Long): List<AuthorEntity>
+    suspend fun getNotSynchronizedAuthors(timestamp: Long, userId: Int): List<AuthorEntity>
 }
