+++ feature/book-info/data/src/commonMain/kotlin/database/LocalBookInfoDataSource.kt
@@ -13,14 +13,14 @@ class LocalBookInfoDataSource(
     private val booksDao = roomDb.booksDao
     private val bookTimestampDao = roomDb.bookTimestampDao
 
-    suspend fun getBookTimestamp(userId: Long) =
+    suspend fun getBookTimestamp(userId: Int) =
         bookTimestampDao.getTimestamp(userId).firstOrNull() ?: createEmptyTimestamp(userId)
 
     suspend fun updateBookTimestamp(bookTimestamp: BookTimestampEntity) {
         bookTimestampDao.insertOrUpdateTimestamp(bookTimestamp)
     }
 
-    suspend fun addOrUpdateBooks(books: List<BookEntity>, userId: Long) {
+    suspend fun addOrUpdateBooks(books: List<BookEntity>, userId: Int) {
         books.forEach { book ->
             val existedBook = booksDao.getBookByBookId(book.bookId, userId = userId).firstOrNull()
             if (existedBook == null) {
@@ -31,18 +31,18 @@ class LocalBookInfoDataSource(
         }
     }
 
-    suspend fun getNotSynchronizedBooks(userId: Long): List<BookEntity> {
+    suspend fun getNotSynchronizedBooks(userId: Int): List<BookEntity> {
         val timestamp = getBookTimestamp(userId)
         return booksDao.getNotSynchronizedBooks(timestamp.thisDeviceTimestamp, userId = userId)
     }
 
-    suspend fun getLocalBookByLocalId(bookLocalId: Long, userId: Long): Flow<List<BookEntity>> =
+    suspend fun getLocalBookByLocalId(bookLocalId: Long, userId: Int): Flow<List<BookEntity>> =
         booksDao.getLocalBookByLocalId(bookLocalId = bookLocalId, userId = userId)
 
-    suspend fun getLocalBookById(bookId: String, userId: Long): Flow<List<BookEntity>> =
+    suspend fun getLocalBookById(bookId: String, userId: Int): Flow<List<BookEntity>> =
         booksDao.getLocalBookById(bookId = bookId, userId = userId)
 
-    suspend fun updateBookAndTime(book: BookEntity, userId: Long): BookEntity {
+    suspend fun updateBookAndTime(book: BookEntity, userId: Int): BookEntity {
         val time = platformInfo.getCurrentTime().timeInMillis
         val existedBook = booksDao.getBookByBookId(book.bookId, userId = userId).firstOrNull()
         if (existedBook == null) {
@@ -63,7 +63,7 @@ class LocalBookInfoDataSource(
         return booksDao.getBookByBookId(book.bookId, userId = userId).first()
     }
 
-    suspend fun updateBookWithoutUpdateTime(book: BookEntity, userId: Long) {
+    suspend fun updateBookWithoutUpdateTime(book: BookEntity, userId: Int) {
         val localId = booksDao.getBookByBookId(book.bookId, userId = userId).firstOrNull()?.localId
         if (localId != null) {
             booksDao.updateBook(book.copy(localId = localId))
@@ -72,7 +72,7 @@ class LocalBookInfoDataSource(
         }
     }
 
-    private suspend fun createEmptyTimestamp(userId: Long): BookTimestampEntity {
+    private suspend fun createEmptyTimestamp(userId: Int): BookTimestampEntity {
         val timestamp = BookTimestampEntity(
             userId = userId,
             otherDevicesTimestamp = 0,
